requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector Store
faiss-cpu>=1.9.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# OpenAI
openai==1.3.0

# Streamlit
streamlit==1.29.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
